requirements2.txt
absl-py>=1.0.0
autopep8>=1.6.0
cachetools>=5.0.0
certifi>=2021.10.8
charset-normalizer>=2.0.12
click>=8.1.2
colorama>=0.4.4
cycler>=0.11.0
Flask>=2.1.1
Flask-SQLAlchemy>=2.5.1
fonttools>=4.32.0
google-auth>=2.6.5
google-auth-oauthlib>=0.4.6
greenlet>=1.1.2
grpcio>=1.45.0
idna>=3.3
imageio>=2.17.0
importlib-metadata>=4.11.3
itsdangerous>=2.1.2
Jinja2>=3.1.1
joblib>=1.1.0
kiwisolver>=1.4.2
Markdown>=3.3.6
MarkupSafe>=2.1.1
matplotlib>=3.5.1
networkx>=2.8
numpy>=1.22.3
oauthlib>=3.2.0
opencv-python>=4.5.5.64
packaging>=21.3
pandas>=1.4.2
Pillow>=9.1.0
protobuf>=3.20.0
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycodestyle>=2.8.0
PyMySQL>=1.0.2
pyparsing>=3.0.8
python-dateutil>=2.8.2
pytz>=2022.1
PyWavelets>=1.3.0
PyYAML>=6.0
requests>=2.27.1
requests-oauthlib>=1.3.1
rsa>=4.8
scikit-image>=0.19.2
scikit-learn>=1.0.2
scipy>=1.8.0
seaborn>=0.11.2
six>=1.16.0
SQLAlchemy>=1.4.35
tensorboard>=2.8.0
tensorboard-data-server>=0.6.1
tensorboard-plugin-wit>=1.8.1
thop>=0.0.31.post2005241907
threadpoolctl>=3.1.0
tifffile>=2022.4.8
toml>=0.10.2
torch>=1.11.0
torchvision>=0.12.0
tqdm>=4.64.0
typing_extensions>=4.2.0
urllib3>=1.26.9
Werkzeug>=2.1.1
yact>=0.5.1
zipp>=3.8.0
